Move logchannel to its own package
