+++ main.go
@@ -1,18 +1,16 @@
 package main
 
 import (
-	"encoding/binary"
 	"fmt"
 	"io"
 	"log"
 	"os"
 	"sync"
-	"syscall"
-	"time"
-
-	"golang.org/x/crypto/ssh"
 
 	gssh "github.com/gliderlabs/ssh"
+	"golang.org/x/crypto/ssh"
+
+	"github.com/arkan/bastion/pkg/logchannel"
 )
 
 func main() {
@@ -67,7 +65,7 @@ func main() {
 }
 
 func proxy(reqs1, reqs2 <-chan *ssh.Request, channel1 ssh.Channel, channel2 ssh.Channel) {
-	wrappedChannel1 := newLoggedChannel(channel1)
+	wrappedChannel1 := logchannel.New(channel1)
 	var closer sync.Once
 	closeFunc := func() {
 		wrappedChannel1.Close()
@@ -113,48 +111,4 @@ func proxy(reqs1, reqs2 <-chan *ssh.Request, channel1 ssh.Channel, channel2 ssh.
 			return
 		}
 	}
-}
-
-type logChannel struct {
-	channel ssh.Channel
-	file    *os.File
-}
-
-func writeTTYRecHeader(fd io.Writer, length int) {
-	t := time.Now()
-
-	tv := syscall.NsecToTimeval(t.UnixNano())
-
-	binary.Write(fd, binary.LittleEndian, int32(tv.Sec))
-	binary.Write(fd, binary.LittleEndian, int32(tv.Usec))
-	binary.Write(fd, binary.LittleEndian, int32(length))
-}
-
-func newLoggedChannel(channel ssh.Channel) *logChannel {
-	f, err := os.OpenFile("session.ttyrec", os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0640)
-	if err != nil {
-		panic(err)
-	}
-
-	return &logChannel{
-		channel: channel,
-		file:    f,
-	}
-}
-
-func (l *logChannel) Read(data []byte) (int, error) {
-	return l.Read(data)
-}
-
-func (l *logChannel) Write(data []byte) (int, error) {
-	writeTTYRecHeader(l.file, len(data))
-	l.file.Write(data)
-
-	return l.channel.Write(data)
-}
-
-func (l *logChannel) Close() error {
-	l.file.Close()
-
-	return l.channel.Close()
 }
